Added some tests and pyproject formulations
