+++ pyproject.toml
@@ -1,0 +1,20 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "rayleigh-py"
+version = "0.1.0"
+description = "A Python package for Rayleigh beam and plate analysis."
+authors = [
+    { name = "Alex Vaughan", email = "acvaughan2000@gmail.com" }
+]
+dependencies = [
+    "numpy>=1.21.0",
+    "scipy>=1.7.0",
+    "matplotlib>=3.4.0"
+]
+
+[tool.setuptools]
+
+packages = ["rayleigh"]
