+++ rayleigh/__init__.py
@@ -1,0 +1,2 @@
+__all__ = ["structs"]
+from . import structs
