+++ rayleigh_py.egg-info/PKG-INFO
@@ -1,0 +1,8 @@
+Metadata-Version: 2.4
+Name: rayleigh-py
+Version: 0.1.0
+Summary: A Python package for Rayleigh beam and plate analysis.
+Author-email: Alex Vaughan <acvaughan2000@gmail.com>
+Requires-Dist: numpy>=1.21.0
+Requires-Dist: scipy>=1.7.0
+Requires-Dist: matplotlib>=3.4.0
