+++ rayleigh_py.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+README.md
+pyproject.toml
+rayleigh/__init__.py
+rayleigh/calculateMatrix.py
+rayleigh/lambda_finder.py
+rayleigh/mainrun.py
+rayleigh/solver.py
+rayleigh/structs.py
+rayleigh_py.egg-info/PKG-INFO
+rayleigh_py.egg-info/SOURCES.txt
+rayleigh_py.egg-info/dependency_links.txt
+rayleigh_py.egg-info/requires.txt
+rayleigh_py.egg-info/top_level.txt
