+++ rayleigh_py.egg-info/requires.txt
@@ -1,0 +1,3 @@
+numpy>=1.21.0
+scipy>=1.7.0
+matplotlib>=3.4.0
